+++ recommender.h
@@ -83,7 +83,7 @@ learn(struct training_set* tset, struct learning_model model);
 *
 */
 double
-estimate_rating_from_factors(int user_index, int item_index, learned_factors_t* lfactors, struct learning_model model);
+estimate_rating_from_factors(size_t user_index, size_t item_index, learned_factors_t* lfactors, struct learning_model model);
 
 
 /*
@@ -102,6 +102,6 @@ estimate_rating_from_factors(int user_index, int item_index, learned_factors_t* 
 *
 */
 recommended_items_t*
-recommend_items(int user_index, int items_number, learned_factors_t* lfactors, training_set_t* tset, learning_model_t model);
+recommend_items(size_t user_index, size_t items_number, learned_factors_t* lfactors, training_set_t* tset, learning_model_t model);
 
 #endif //RECOMMANDER_H
